Extract TeamRunLiveModel and collapse the team-run dispatcher cases
The live team run's presentation — the dispatcher activity card, per-agent
activity rows, the pending dispatch plan, and team metering — moves behind
a feature model fed one event at a time through apply(_:_:). Fourteen
dispatcher cases collapse to one-line routing: the dispatcher keeps the
run's identity and state scalars plus task-conversation bookkeeping, and
hands each event's presentation slice to the model. Session restore and
reset write activities through intent-named seams instead of raw state,
and the shared metering arithmetic that was repeated across six cases
becomes one helper with the delegated/metered distinction the events
historically carried.

New tests cover the dispatcher card lifecycle, agent-spawn dedup by node,
job completion carrying identity from the started row, new-run resets,
the two-condition dispatch-approval gate, and the bridge.
